refactor & rename shared components for accelerator report targets views
